refactor: 상품상세페이지 컴포넌트화 작업
